Rename: ModaFrame.type.ts -> ModalFrame.type.ts 파일명 오타 수정
