+++ src/components/shared/Modal/ModalFrame.type.ts
